feat: 파일 관리 구현

+++ src/main/java/com/sprint/hrbank/application/employee/required/EmployeeRepository.java
@@ -13,5 +13,5 @@ public interface EmployeeRepository extends JpaRepository<Employee, Long>, Emplo
 
   Optional<Employee> findByEmployeeNumber(String employeeNumber);
 
-  Integer countEmployeesByDepartment_Id(Long departmentId);
+  Long countEmployeesByDepartment_Id(Long departmentId);
 }
